feat(toolbar): add transparent toolbar for plans view
The changes introduce a new `TransparentToolbar` component that is
displayed when the user is in the "plans" view. This toolbar has a
transparent background and a border-bottom to match the design of the
plans view. The main toolbar logic is updated to render the
`TransparentToolbar` when the user is in the plans view.