feat(timer): add click-outside handling for editing mode
Adds a new effect that listens for mouse down events outside the
input element when the timer is in editing mode. When a click
outside the input is detected, the editing mode is exited.

This improves the user experience by providing a more intuitive way
to exit the editing mode without having to manually click the
"Done" button.