feat(tiptap): add custom list keymap
Adds a custom list keymap that extends the default list keymap
provided by the Tiptap library. The custom keymap handles the
Enter and Backspace keys to provide a better user experience
when working with list items.

The Enter key now lifts the list item if the current list item
is empty, allowing the user to easily exit the list. The Backspace
key also lifts the list item if the current list item is empty and
the cursor is at the beginning of the list item, allowing the user
to easily remove the list item.