feat(windows): add new Finder window type
The changes in this commit introduce a new window type called "Finder" to the `HyprWindow` type in the `bindings.gen.ts` file. This change is made to accommodate the addition of a new Finder-like functionality in the application.

Additionally, the corresponding changes are made in the `routeTree.gen.ts` file to add a new route for the Finder functionality, replacing the previously existing Calendar route.